ZeitauswahlStundeOrBlock als componente verpackt.
